# docs/openapi.json
{
  "components": {
    "schemas": {
      "Body_ingest_v1_ingest_post": {
        "properties": {
          "entity_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Id"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "logical_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Groups versions of the same file",
            "title": "Logical Name"
          },
          "module": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Module"
          },
          "product_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Id"
          },
          "tenant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_ingest_v1_ingest_post",
        "type": "object"
      },
      "DevTokenRequest": {
        "properties": {
          "persona": {
            "title": "Persona",
            "type": "string"
          }
        },
        "required": [
          "persona"
        ],
        "title": "DevTokenRequest",
        "type": "object"
      },
      "DevTokenResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in_min": {
            "title": "Expires In Min",
            "type": "integer"
          },
          "persona": {
            "title": "Persona",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "expires_in_min",
          "persona"
        ],
        "title": "DevTokenResponse",
        "type": "object"
      },
      "ExportFilters": {
        "properties": {
          "date_from": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date From"
          },
          "date_to": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date To"
          },
          "employee_id": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Id"
          },
          "entity_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Id"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "present",
                  "absent",
                  "leave",
                  "holiday",
                  "wfh",
                  "half_day",
                  "unknown",
                  "conflict"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "title": "ExportFilters",
        "type": "object"
      },
      "ExportRequest": {
        "properties": {
          "filters": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExportFilters"
              },
              {
                "type": "null"
              }
            ]
          },
          "format": {
            "default": "json",
            "enum": [
              "json",
              "xlsx",
              "pdf"
            ],
            "title": "Format",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "source": {
            "default": "records",
            "enum": [
              "records",
              "query"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "title": "ExportRequest",
        "type": "object"
      },
      "FeedbackRequest": {
        "properties": {
          "agent_response": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Response"
          },
          "approve": {
            "default": true,
            "title": "Approve",
            "type": "boolean"
          },
          "feedback": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Feedback",
            "type": "string"
          },
          "ideal_final_output": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Ideal Final Output",
            "type": "string"
          },
          "original_request_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Original Request Id",
            "type": "string"
          },
          "question": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          }
        },
        "required": [
          "original_request_id",
          "feedback",
          "ideal_final_output"
        ],
        "title": "FeedbackRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "QueryFilters": {
        "properties": {
          "date_from": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date From"
          },
          "date_to": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date To"
          },
          "employee_id": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Id"
          },
          "entity_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Id"
          }
        },
        "title": "QueryFilters",
        "type": "object"
      },
      "QueryRequest": {
        "properties": {
          "filters": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QueryFilters"
              },
              {
                "type": "null"
              }
            ]
          },
          "include_debug": {
            "default": false,
            "title": "Include Debug",
            "type": "boolean"
          },
          "question": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "QueryRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "JWT from /v1/auth/dev-token (dev only)",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Enterprise RAG attendance microservice: ingestion, normalization, isolated retrieval, grounded answers with citations, feedback and export.",
    "title": "Attendance Intelligence API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/auth/dev-token": {
      "post": {
        "operationId": "dev_token_v1_auth_dev_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Issue a dev token",
        "tags": [
          "auth (dev only)"
        ]
      }
    },
...
    "/v1/export": {
      "post": {
        "operationId": "export_v1_export_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export permitted records (or a stored answer's evidence) as JSON, XLSX or PDF",
        "tags": [
          "export"
        ]
      }
    },
    "/v1/feedback": {
      "get": {
        "operationId": "list_feedback_v1_feedback_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Feedback V1 Feedback Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Feedback examples in your scope",
        "tags": [
          "feedback"
        ]
      },
      "post": {
        "operationId": "submit_v1_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Submit V1 Feedback Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit a correction + ideal output for a previous answer",
        "tags": [
          "feedback"
        ]
      }
    },
...
    "/v1/ingest": {
      "post": {
        "operationId": "ingest_v1_ingest_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_v1_ingest_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Ingest V1 Ingest Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "description": "Queued for background processing"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload an attendance file (CSV, XLSX; more formats in later steps)",
        "tags": [
          "ingestion"
        ]
      }
    },
...
    "/v1/query": {
      "post": {
        "operationId": "query_v1_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Query V1 Query Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ask an attendance question (answers only from permitted data)",
        "tags": [
          "query"
        ]
      }
    },
    "/v1/query/{request_id}": {
      "get": {
        "operationId": "get_query_v1_query__request_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Query V1 Query  Request Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "A stored answer (only within your scope)",
        "tags": [
          "query"
        ]
      }
    }
  }
}